untested initial seek integration
